Rename driver::register to driver::buffers
